move classes to separate file
